Move collapsible code into its own js file
