+++ wagtail/admin/templates/wagtailadmin/home.html
@@ -35,6 +35,6 @@
 
 {% block extra_js %}
 {{ block.super }}
-<script src="{% versioned_static 'wagtailadmin/js/page-editor.js' %}"></script>
+<script src="{% versioned_static 'wagtailadmin/js/collapsible.js' %}"></script>
 {{ media.js }}
 {% endblock %}
